chore: moves navigation files to new navigation folder
