+++ App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import store from './client/store';
 
-import RootContainer from './RootContainer';
+import RootContainer from './client/navigation/RootContainer';
 
 const App = () => {
   return (
